fix resources menu and space above v2 projects
